requirements.txt
fastapi
uvicorn
python-docx
python-multipart
python-dotenv
openai
markdown
pydantic
pydantic_settings
httpx